Integrated trakt.tv API
We are now back to the previous UI glory that came before the original project shutdown. This commit has a bit of hackish nature to it, but after 3 hours of staring at my screen, this was the only solution I could think of (regarding the Backbone models). Everything looks sweet now :smile: YAY!! RE: Issue #17